Improved definition files
